Separated graphics and physics modules (SOC)
