+++ sim_game/game.py
@@ -6,7 +6,8 @@ __authors__    = ["Ole Herman Schumacher Elgesem"]
 __license__    = "MIT"
 # This file is subject to the terms and conditions defined in 'LICENSE'
 
-from graphics import SpriteObject, TextObject, GraphicsRectangle, Renderer
+from graphics import GraphicsRectangle, Renderer
+from physics import SpriteObject, TextObject
 from robot import Robot
 
 import random
